remove global `_analyzer` variable and refactor
