+++ templates/pubspecs/tools/serverpod_cli/pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
     path: ../../packages/serverpod_shared
   serverpod_service_client: #^VERSION
     path: ../../packages/serverpod_service_client
+  async: ^2.11.0
 
 dev_dependencies:
   lints: ^2.0.0
