+++ tools/serverpod_cli/bin/serverpod_cli.dart
@@ -5,6 +5,7 @@ import 'package:args/args.dart';
 import 'package:colorize/colorize.dart';
 
 import 'package:serverpod_cli/src/analytics/analytics.dart';
+import 'package:serverpod_cli/src/analyzer/dart/protocol_analyzer.dart';
 import 'package:serverpod_cli/src/create/create.dart';
 import 'package:serverpod_cli/src/downloads/resource_manager.dart';
 import 'package:serverpod_cli/src/generated/version.dart';
@@ -204,13 +205,19 @@ Future<void> _main(List<String> args) async {
         return;
       }
 
+      var analyzer = ProtocolDartFileAnalyzer(config.endpointsSourcePath);
       await performGenerate(
         verbose: verbose,
         config: config,
+        analyzer: analyzer,
       );
       if (watch) {
         print('Initial code generation complete. Listening for changes.');
-        performGenerateContinuously(verbose, config);
+        performGenerateContinuously(
+          verbose,
+          config,
+          analyzer: analyzer,
+        );
       } else {
         print('Done.');
       }
