+++ tools/serverpod_cli/lib/src/analyzer/dart/protocol_analyzer.dart
@@ -9,11 +9,11 @@ import 'package:analyzer/diagnostic/diagnostic.dart';
 import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:source_span/source_span.dart';
 
-import '../util/subdirectory_extraction.dart';
-import 'config.dart';
-import '../analyzer/dart/definitions.dart';
-import 'code_analysis_collector.dart';
-import 'types.dart';
+import '../../util/subdirectory_extraction.dart';
+import '../../generator/config.dart';
+import 'definitions.dart';
+import '../../generator/code_analysis_collector.dart';
+import '../../generator/types.dart';
 
 const _excludedMethodNameSet = {
   'streamOpened',
@@ -24,59 +24,23 @@ const _excludedMethodNameSet = {
   'getUserObject',
 };
 
-ProtocolAnalyzer? _analyzer;
-
-Future<ProtocolDefinition> performAnalyzeServerCode({
-  required bool verbose,
-  required CodeAnalysisCollector collector,
-  bool requestNewAnalyzer = true,
-  required Set<String> changedFiles,
-  required GeneratorConfig config,
-}) async {
-  // Invalidate the old analyzer if requested to do so.
-  if (requestNewAnalyzer) {
-    _analyzer = null;
-  }
-
-  // Attempt to use old analyzer or create a new one if no one exists.
-  if (_analyzer == null) {
-    _analyzer = ProtocolAnalyzer(config.endpointsSourcePath);
-  } else if (changedFiles.isNotEmpty) {
-    // Make sure that the analyzer is up-to-date with recent changes.
-    var contexts = _analyzer!.collection.contexts;
-    for (var context in contexts) {
-      for (var changedFile in changedFiles) {
-        var file = File(changedFile);
-        context.changeFile(file.absolute.path);
-      }
-      await context.applyPendingFileChanges();
-    }
-  }
-
-  // Perform the code analysis.
-  return await _analyzer!.analyze(
-    verbose: verbose,
-    collector: collector,
-  );
-}
-
-Future<List<String>> performAnalysisGetSevereErrors(
-    GeneratorConfig config) async {
-  _analyzer = ProtocolAnalyzer(config.endpointsSourcePath);
-  return await _analyzer!.getErrors();
-}
-
-class ProtocolAnalyzer {
+/// Analyzes dart files for the protocol specification.
+class ProtocolDartFileAnalyzer {
   final Directory endpointDirectory;
   late AnalysisContextCollection collection;
 
-  ProtocolAnalyzer(String filePath) : endpointDirectory = Directory(filePath) {
+  /// Create a new [ProtocolDartFileAnalyzer], analyzing
+  /// all dart files in [filePath].
+  //TODO: Make ProtocolDartFileAnalyzer testable
+  ProtocolDartFileAnalyzer(String filePath)
+      : endpointDirectory = Directory(filePath) {
     collection = AnalysisContextCollection(
       includedPaths: [endpointDirectory.absolute.path],
       resourceProvider: PhysicalResourceProvider.INSTANCE,
     );
   }
 
+  /// Get all errors in the analyzed files.
   Future<List<String>> getErrors() async {
     var errorMessages = <String>[];
 
@@ -99,10 +63,24 @@ class ProtocolAnalyzer {
     return errorMessages;
   }
 
+  /// Analyze all files in the [endpointDirectory].
+  /// Use [changedFiles] to mark files, that need reloading.
   Future<ProtocolDefinition> analyze({
     required bool verbose,
     required CodeAnalysisCollector collector,
+    Set<String>? changedFiles,
   }) async {
+    if (changedFiles != null) {
+      // Make sure that the analyzer is up-to-date with recent changes.
+      for (var context in collection.contexts) {
+        for (var changedFile in changedFiles) {
+          var file = File(changedFile);
+          context.changeFile(file.absolute.path);
+        }
+        await context.applyPendingFileChanges();
+      }
+    }
+
     var endpointDefs = <EndpointDefinition>[];
     var filePaths = <String>[];
