+++ tools/serverpod_cli/pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
     path: ../../packages/serverpod_shared
   serverpod_service_client: #^1.0.0
     path: ../../packages/serverpod_service_client
+  async: ^2.11.0
 
 dev_dependencies:
   lints: ^2.0.0
